Fix quick search

+++ app/src/main/java/com/hippo/ehviewer/ui/scene/GalleryListScene.java
@@ -1533,8 +1533,8 @@ public final class GalleryListScene extends BaseScene
                     setState(STATE_NORMAL);
                     closeDrawer(Gravity.RIGHT);
                 });
-                holder.option.setOnClickListener(v -> {
-                    PopupMenu popupMenu = new PopupMenu(getContext2(), holder.option);
+                holder.itemView.setOnLongClickListener(v -> {
+                    PopupMenu popupMenu = new PopupMenu(requireContext(), holder.option);
                     popupMenu.inflate(R.menu.quicksearch_option);
                     popupMenu.show();
                     popupMenu.setOnMenuItemClickListener(new PopupMenu.OnMenuItemClickListener() {
@@ -1543,16 +1543,16 @@ public final class GalleryListScene extends BaseScene
 
                         @Override
                         public boolean onMenuItemClick(MenuItem item) {
-                            switch (item.getItemId()) {
-                                case R.id.menu_qs_remove:
-                                    EhDB.deleteQuickSearch(quickSearch);
-                                    mQuickSearchList.remove(position);
-                                    notifyDataSetChanged();
-                                    break;
+                            if (item.getItemId() == R.id.menu_qs_remove) {
+                                EhDB.deleteQuickSearch(quickSearch);
+                                mQuickSearchList.remove(position);
+                                notifyDataSetChanged();
+                                return true;
                             }
                             return false;
                         }
                     });
+                    return true;
                 });
             }
         }
@@ -1569,7 +1569,7 @@ public final class GalleryListScene extends BaseScene
 
         @Override
         public boolean onCheckCanStartDrag(@NonNull QsDrawerHolder holder, int position, int x, int y) {
-            return true;
+            return x > holder.option.getX() && y > holder.option.getY();
         }
 
         @Override
